s3validation: Pass token during ec2 role auth (PROJQUAY-2983)
When the access and secret key aren't passed during s3 storage validation it will fall back to using ec2 role auth. This will fail with `The AWS Access Key Id you provided does not exist in our records.`. This fix passes the session token so that the auth to s3 can succeed.